requirements.txt
Adafruit_GPIO == 1.0.3
Adafruit_SSD1306 == 1.6.2
Flask == 3.1.0
Pillow == 11.1.0
RPi.BME280 == 0.2.4
RPi.GPIO == 0.7.1
smbus == 1.1.post2
